refactor: move RichText class from model to types directory and update imports
