+++ packages/core/src/model/view/LikeC4ViewModel.ts
@@ -8,11 +8,11 @@ import type {
   ViewWithType,
 } from '../../types'
 import { _stage, _type } from '../../types'
+import { type RichTextOrEmpty, RichText } from '../../types'
 import type * as aux from '../../types/aux'
 import { DefaultMap, ifind, nonNullable } from '../../utils'
 import type { ElementModel } from '../ElementModel'
 import type { LikeC4Model } from '../LikeC4Model'
-import { type RichTextOrEmpty, RichText } from '../RichText'
 import {
   type $View,
   type EdgeOrId,
