+++ packages/layouts/src/graphviz/dot-labels.ts
@@ -9,8 +9,8 @@ import {
   defaultTheme,
   defaultTheme as Theme,
   nonexhaustive,
+  RichText,
 } from '@likec4/core'
-import { RichText } from '@likec4/core/model'
 import { identity, isDefined, isTruthy } from 'remeda'
 import wordWrap from 'word-wrap'
 import { IconSizePoints, pxToPoints } from './utils'
@@ -28,14 +28,14 @@ export function wrap(text: string, {
   maxchars,
   maxLines,
   sanitize: escape = identity(),
-}: WrapOptions) {
+}: WrapOptions): string[] {
   let lines = wordWrap(text, {
     width: maxchars,
     indent: '',
     escape,
   }).split('\n')
   if (isDefined(maxLines) && maxLines > 0 && lines.length > maxLines) {
-    lines = lines.slice(0, maxLines - 1)
+    lines = lines.slice(0, maxLines)
   }
   return lines
 }
@@ -121,11 +121,11 @@ export function nodeLabel(
         }),
       )
     }
-    const description = node.description ? RichText.from(node.description) : undefined
+    const description = RichText.from(node.description).text
     if (description) {
       lines.push(
         wrapWithFont({
-          text: description.text,
+          text: description,
           fontsize: Math.ceil(fontSize(sizes.text) * 0.75),
           maxchars: hasIcon ? 35 : 45,
           maxLines: isSmOrXs ? 3 : 5,
@@ -134,7 +134,7 @@ export function nodeLabel(
       )
     }
   }
-  if (lines.length === 1 && !hasIcon) {
+  if (lines.length === 1 && hasIcon === false) {
     return `<${lines[0]}>`
   }
